+++ src/Workspaces/Core/Portable/EmbeddedLanguages/LanguageServices/LanguageCommentDetector.cs
@@ -4,54 +4,11 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
-using System.Text.RegularExpressions;
 
 namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
 {
-    /// <summary>
-    /// Helps match patterns of the form: language=name,option1,option2,option3
-    /// <para/>
-    /// All matching is case insensitive, with spaces allowed between the punctuation. Option values are returned as strings.
-    /// <para/>
-    /// Option names are the values from the TOptions enum.
-    /// </summary>
-    internal struct EmbeddedLanguageCommentDetector
-    {
-        private readonly Regex _regex;
-
-        public EmbeddedLanguageCommentDetector(string[] identifiers)
-        {
-            var namePortion = string.Join("|", identifiers.Select(n => $"({Regex.Escape(n)})"));
-            _regex = new Regex($@"^((//)|(')|(/\*))\s*lang(uage)?\s*=\s*(?<identifier>{namePortion})\b((\s*,\s*)(?<option>[a-zA-Z]+))*",
-                RegexOptions.ExplicitCapture | RegexOptions.IgnoreCase | RegexOptions.Compiled);
-        }
-
-        public bool TryMatch(
-            string text,
-            [NotNullWhen(true)] out string? identifier,
-            [NotNullWhen(true)] out IEnumerable<string>? options)
-        {
-            var match = _regex.Match(text);
-            identifier = null;
-            options = null;
-            if (!match.Success)
-                return false;
-
-            identifier = match.Groups["identifier"].Value;
-
-            var optionGroup = match.Groups["option"];
-#if NETCOREAPP
-            options = optionGroup.Captures.Select(c => c.Value);
-#else
-            options = optionGroup.Captures.OfType<Capture>().Select(c => c.Value);
-#endif
-            return true;
-        }
-    }
-
     /// <summary>
     /// Helps match patterns of the form: language=name,option1,option2,option3
     /// <para/>
@@ -80,6 +37,13 @@ namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
             options = default;
             if (!_detector.TryMatch(text, out _, out var captures))
                 return false;
+
+            return TryGetOptions(captures, out options);
+        }
+
+        public static bool TryGetOptions(IEnumerable<string> captures, out TOptions options)
+        {
+            options = default;
 
             foreach (var capture in captures)
             {
